Add BKTClient (#87)
* feat: add BKTUser and UserHolder

* chore: make builder method public

* refactor: remove json model namespace

* feat: add di classes

* feat: add foreground scheduler

* test: add tests for foreground scheduler

* fix: remove duplicate constant values

* fix: remove warnings

* fix: the codes pointed out

* feat: add bkt client

* test: add evaluation tests

* test: add bkt client tests

* fix: pointed out

* fix: change json type to dictionary

* fix: pointed out case

* fix: removed unnecessary case

+++ Bucketeer2/Sources/Internal/Model/MetricsEventData.swift
@@ -13,7 +13,7 @@ enum MetricsEventData: Hashable {
 
     struct GetEvaluationSize: Codable, Hashable {
         let labels: [String: String]
-        let size_byte: Int
+        let size_byte: Int64
     }
 
     struct TimeoutErrorCount: Codable, Hashable {
